Add progress for ActionCodeSettingsBuilder, TokenVerifier
